Fix build failures
Fix the following similar build failures, some of the build failure
reproduced with arm64/riscv64, some of the build reproduced with 32bit
x86/arm.

dlt_daemon_common.c:2785:26: error: conversion to 'int8_t' {aka 'signed char'} from 'char' may change the sign of the result [-Werror=sign-conversion]

src/tests/dlt-test-multi-process-client-v2.c:302:49: error: format '%lld' expects argument of type 'long long int', but argument 2 has type 'time_t' {aka 'long int'} [-Werror=format=]
  302 |     printf(" Time running (seconds)        : %lld\n", time(NULL) - stats.first_message_time);
      |                                              ~~~^     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
      |                                                 |                |
      |                                                 long long int    time_t {aka long int}
      |                                              %ld

src/lib/dlt_user.c:5109:60: error: format '%lld' expects argument of type 'long long int', but argument 4 has type '__off_t' {aka 'long int'} [-Werror=format=]
 5109 |             dlt_vlog(LOG_DEBUG, "%s: Current file size=[%lld]\n", __func__,
      |                                                         ~~~^
      |                                                            |
      |                                                            long long int
      |                                                         %ld
 5110 |                      st.st_size);
      |                      ~~~~~~~~~~
      |                        |
      |                        __off_t {aka long int}

+++ src/daemon/dlt-daemon.c
@@ -457,7 +457,7 @@ int option_file_parser(DltDaemonLocal *daemon_local)
 
     if (n < 0 || (size_t)n > sizeof(daemon_local->flags.loggingFilename)) {
         dlt_vlog(LOG_WARNING, "%s: snprintf truncation/error(%ld) %s\n",
-                __func__, n, daemon_local->flags.loggingFilename);
+                __func__, (long int)n, daemon_local->flags.loggingFilename);
     }
     daemon_local->flags.enableLoggingFileLimit = false;
     daemon_local->flags.loggingFileSize = 250000;
@@ -2501,7 +2501,7 @@ void dlt_daemon_exit_trigger()
     n = snprintf(tmp, DLT_PATH_MAX, "%s/dlt", dltFifoBaseDir);
     if (n < 0 || (size_t)n > DLT_PATH_MAX) {
         dlt_vlog(LOG_WARNING, "%s: snprintf truncation/error(%ld) %s\n",
-                __func__, n, tmp);
+                __func__, (long int)n, tmp);
     }
 
     (void)unlink(tmp);
@@ -2878,8 +2878,8 @@ int dlt_daemon_log_internal(DltDaemon *daemon, DltDaemonLocal *daemon_local,
         dlt_daemon_client_send(DLT_DAEMON_SEND_TO_ALL, daemon,daemon_local,
                             msg.headerbuffer, sizeof(DltStorageHeader),
                             msg.headerbuffer + sizeof(DltStorageHeader),
-                            (size_t)(msg.headersize - (int32_t)sizeof(DltStorageHeader)),
-                            msg.databuffer, (size_t)msg.datasize, verbose);
+                            (int)(msg.headersize - (int32_t)sizeof(DltStorageHeader)),
+                            msg.databuffer, (int)msg.datasize, verbose);
 
         free(msg.databuffer);
     }else {
